WT-15249 Add disagg conflict sub level error code and fix the conflict in test_timestamp26 (#12356)
When log is enabled we are expected to ignore
`write_timestamp_usage=never`, but since disagg is on and it overwrites
log to be disabled, we don't ignore `write_timestamp_usage=never` which
causes the failure.

This PR is to add a `disagg conflict` sub level error code when disagg
is on and `write_timestamp_usage=never `as they are not compatible, and
fixed the bug in test_timestamp26.