WT-14904 Fix building internal page delta for a previously split child (#12317)
We have already disabled building internal page deltas if any child page
split in the most recent reconciliation. However, it is not enough. We
can have a child that was split in reconciliation 1 but not split in
reconciliation 2. If we write the split page in its parent's previous
reconciliation, we will unintentionally not deleting those split keys in
the second reconciliation if we write a delta. Therefore, we should stop
building the internal page delta if we see any child page that is
written as multiple keys in the previous internal page reconciliation.

Added a python test test_layered32.py

---------

Co-authored-by: Ravi Prakash Giri <ravi.giri@mongodb.com>