WT-15389 Fix saved update lists with deletes are lost when building a delta (#12369)
When we have deletes at the start of the page, we may trigger split when
we still have nothing to write. This will implicitly discard the saved
updates with deletes to build a future delta, leading to deleted data
reappear in the future.

Forbid to split when there is still nothing to write. Only use the saved
update with an on-page value for split size calculation.