WT-15300 Set backlink LSN in __wt_disagg_put_meta & add more verification (#12362)
Metadata pages in the special table 1 now set the backlink LSN.

As a part of the testing, I added a verification routine to `plh_put`
that verifies that the base and backlink LSNs are set correctly, but I
had to turn off a part of the checks because PALM does not yet handle
abandoned checkpoints, so tests `test_layered15` and `test_layered27` do
not pass with it enabled. All other tests pass.

I also fixed a few small things. Notably, I removed a space before the
line number in `PALM_KV_RET` and `PALM_KV_ERR` to allow VS Code to
recognize them as locations within the file.