WT-15438 Do not preserve checkpoint_meta and last_materialized_lsn in the connection config (#12386)
We should not preserve certain disaggregated config parameters in the
connection config after `WT_CONNECTION::reconfigure` because:
* Preserving `checkpoint_meta` could cause an unrelated `reconfigure`
call to pick up the same checkpoint again, which is at best wasteful. At
the same time, we cannot ignore the direction to _not_ pick up a
checkpoint if requested, as, for example, `test_layered38` depends on
that behavior.
* Preserving `last_materialized_lsn` could cause problems, because there
are ways to update it other than `reconfigure`.