WT-15148 Use conditional variable to sync threads (#12366)
The changes introduce new conditional variables to the timestamp_abort
test: they are needed to avoid the timeout issue faced. Without them,
the checkpoint thread may run without the test being ready for it. Each
checkpoint run can be slow due to the pressure on the system and lead to
timeout issues. With the condition variables in place, the checkpoint
thread only starts when the test is ready which greatly reduces the
chances to reach the timeout limit.

---------

Co-authored-by: Albert Song <albert.song@mongodb.com>
Co-authored-by: Etienne Petrel <77254506+etienneptl@users.noreply.github.com>