WT-15409 Fix reverse compare assertion (#12389)
The collator and reversed table is working correctly. Because the btree
is reversed and the collator is reversed the expectation is still
correct around cmp <= 0. The problem is we check cmp <= 0 when the
original key is not set. In the normal case this is not an issue because
we expect an empty key to be less than any key. But in reverse case an
empty key is greater than any key, causing the assertion failure.