WT-15254 Add disagg leader hooks to PR testing (#12344)
This is, unfortunately, a bit more difficult than just adding the
`--batch` flag. The `pytest_parallel` script has a bug where `wait -n`
returns, but not for a job. We then can't find a process when calling
`waitone` and it all falls over.

Rather than hunting this down and fixing `pytest_parallel`, I rolled the
block-list functionality into the existing `unit test` function. The
only user left of `unit test parallel` is now some TSAN testing.

There's a link to a patch build on the ticket. One test failed under
TSAN, but it's not because of this change. The maximum task duration for
that patch is 19 minutes, which matches the existing bucketed Python
tests.