WT-15211 incremental app thread eviction (#12397)
This is the same change that was reverted with 474cfca, but with the
feature flag turned off by default. The reverted branch is #12303. The
test format failure in wt-perf are related to an interaction with this
change and precise checkpoints being turned on, that resulted in more
pages being unable to be evicted.

We will tune and test these changes in a separate branch from develop.
See the WT ticket for more details