WT-15441 Fix __clayered_compare failure (#12395)
When performing WT_CURSOR::compare, we should not be using the uri but
instead the internal_uri.

The `cursor->internal_uri` is an internal representation is always set
for a cursor. It's used by various internal cursor code that requires
the internal uri variable. This should be used for public purposes such
as WT_CURSOR::compare.

The `cursor->uri` is the application-visible URI that represents what
the user requested, and thus accurately reflects the correct URI state.

While figuring this out, I also went through triaged all python test
truncate tests and test/format tests and created tickets for python
tests that fail. As a result, evergreen CI should now be running python
truncate tests