WT-15426 Keep a log of errors to pinpoint the origin of the error (#12380)
Keep a "stack trace" of errors—called the error log—processed by
`WT_ERR`, `WT_RET`, and their variants, so that we can easily pinpoint
origins of errors. The error log is enabled only in the diagnostic mode
(to avoid performance penalty for production), maintained per thread,
and cleared on API entry. The error log can be printed via a new
(undocumented) connection function. It is currently also being printed
when reconfiguring disagg fails (we could remove this if we'd like).