WT-15531 Fix error log handling in test_layered31 (#12421)
`test_layered31` does not currently work in the release mode, because it
expects the error log, but the error log is not available in the release
mode.

The solution is two-fold:
* `assertRaisesException` now ignores the file/line number log records,
if they are present, and it also ignores the log record that repeat the
exception, if present.
* `test_layered31` calls `dump_error_log` with
`self.expectedStderrPattern('WT_VERB_ERROR_RETURNS')` only in the
diagnostic mode.