WT-15548 Disable checkpoint_cleanup config in test/format if all related options are off (#12430)
This is most of the WT-15375 PR, pulled across for easier backporting.

It includes the compatibility test script changes that 8.2 needs in
order to generate valid configs for older branches, as well as the
change in the title.

It does not include the dirty restart fixes, which only need to be on
develop.