WT-15375 Dirty restart test compatibility fixes (#12347)
This is the simplest way to fix the test failure.

However, longer term I think we could revise what "backwards compatible"
means -- 4.4 is starting to fall off the lists in `versions.sh`, so I
suspect a bunch of the currently turned-off stuff can be removed from
both (1) this list, and (2) the list of `BC_CHECK` calls in
`format_config.c`.