WT-15494 Fix resolve prepare operations when claiming a pending prepared transaction (#12426)
When claiming a pending prepared transaction, we transfer the
modification ops and assign a new transaction id. However, the WT_UPDATE
object is not saved in the pending prepared map so it does not get the
new txn id, so its time window's transaction id is reset to 0. This led
to issue when commit/rollback, the prepared ops are not resolved
properly because the txn id of the update and transactions don't match.

This PR fixes the issue by checking the state of the prepared updates
instead of relying on txn id when resolving prepared update chain.