WT-15485 Configuration to control app thread eviction based on minimum cache fill ratio (#12404)
This configuration parameter controls the involvement of the application
thread in the eviction whenever the cache fill ratio is greater than the
minimum cache fill ratio that is configured.

With this configuration, even when the dirty or update triggers are met,
the application threads are not involved in the eviction if the minimum
cache fill ratio is not met.