WT-15291 Fix layered cursors reset in error cases + minor improvements (#12425)
The expectation of the cursors API is that cursor should be left
unpositioned in case of returning an error (including NOT_FOUND). That
is not always true for the current logic in layered cursors. This commit
fixes all the found cases where some logic that could lead to unexpected
consequences (e.g.for next() and prev() that could make decisions based
on the previous states) is left defined after the error happens.

This change also includes a couple of minor improvements such as:
- Code readability improvements
- Removing redundant if branches (replacing them with asserts)
- Removing WT_CLAYERED_MULTIPLE redundant flag that is never read