WT-15627 Enable output of compile commands during generation (#12486)
Enable output of compile commands during generation.

When this option is enabled, CMake generates a compile_commands.json
file containing the exact compiler calls for all translation units of
the project in machine-readable form.
This helps VSCode with code navigation, and allows developers to examine
actual compiler flags to be used during build.