WT-15488 test_verify.py fails with mismatch in page IDs from PALM and btree walk (#12492)
The memory allocation for the page_id was incorrect before. With the
macro __wt_realloc_def, the capacity it takes in is in bytes, and the
new capacity is the increment of array size. With the previous
comparison, capacity would be 8 and we never correctly allocate memory
for the array before index 8 due to this check. Therefore we need to
compare `(capacity_in_bytes / sizeof(uint64_t))` instead with
`num_pages_found_in_btree`, so they would both be number of items
tracked/allocated, instead of bytes of memory.