WT-15519 Abort when seeing OOO keys in __verify_row_key_order_check (#12496)
When doing row key order verification, we only throw an error if we find
an OOO key. However this cause the stack trace to lose critical
information for debugging. This PR makes it abort on error if a
diagnostic build is enabled to help with debugging when this issue
happens again.