WT-15623 Add --print-output option for run.py (#12515)
Currently, by default, run.py intercepts stdout and stderr, redirecting
them to separate files to check for unexpected output later. The goal of
this ticket is to create an option that also allows duplicating the
output to the normal stdout/stderr, which can be useful for debugging.

**This is the second PR, [the first
one](https://github.com/wiredtiger/wiredtiger/pull/12477) was reverted
because of a CI failure caused in compatibility tests. The last commit
fixes it.**