WT-15629 Disable building delta for the first internal key (#12551)
This change disables the construction of an internal page delta when the
first key on the page is updated. We've observed that this can lead to
out-of-order issues, both with the default collator and a custom
collator. The specific reason it fails with the default collator remains
unclear. However, when using a custom collator, truncating the first key
on the page can cause it to become out of order relative to other keys.
As a result, merging the delta may place the first key incorrectly.

To avoid the complexity of handling all the edge cases related to the
first key, we’ve decided to disable delta building for internal pages
whenever the first key is updated.

---------

Co-authored-by: zunyi.liu <zunyi.liu@mongodb.com>
Co-authored-by: Albert <whoisong@gmail.com>
Co-authored-by: Albert Song <albert.song@mongodb.com>
Co-authored-by: Alex Blekhman <alexander.blekhman@mongodb.com>
Co-authored-by: Zunyi Liu <43444919+zunyiliu@users.noreply.github.com>
Co-authored-by: Monica Ng <monica.ng@mongodb.com>
Co-authored-by: Ravi Giri <56813441+raviprakashgiri29@users.noreply.github.com>
Co-authored-by: Hari Babu Kommi <haribabu.kommi@mongodb.com>