WT-15560 Fix failing assertion when unpacking a prepared tombstone (#12566)
This PR fixes an edge case when unpacking a cell with stop_txn ==
start_txn == 0. Consider this scenario with precise_checkpoint=ON and
preserve_prepared=ON:

- A prepared tombstone with a globally visible start is saved to disk.
At this point, start_txn = 0 and stop_txn != 0
- In recovery, we unpack this cell and reset all the cell's start & stop
txn to 0.
- Prepare discover cursor claims the tombstone update. However we don't
create a new transaction for this claim prepare work, so `stop_txn`
never got assigned a new txn id.
- If stable ts doesn't move and we call a checkpoint, the tombstone will
still be saved as prepared (because stop ts is not stable). It will be
saved as prepared and stop_txn = 0 == start_txn, which fails a validity
rule when unpack again.

This PR fixes that by relaxing the cell unpacking validity rule.