WT-15680 Disable layered table verification in Mongo (#12571)
This PR disables verification of layered tables when running as part of
MongoDB while keeping it enabled for standalone WT builds.

It's a temporary solution to silence test failure errors until proper
verification with deltas is implemented.

Implementing partial verification of tables without deltas would be too
complex to justify as a short-term workaround.