WT-13454 Support compiled configurations for WT_SESSION reconfigure (#12460)
This PR adds compiled configuration support for the
`WT_SESSION::reconfigure` API.
Correctness test is implemented by extending
`test/csuite/config/main.c`.
No benchmark/perf test is added at the WT level, as the mongo level
testing is more meaningful and is expected to be covered by the
dependency server ticket SERVER-93982.

You may notice there are 2 prefetch API configuration keys and structs
defined in `src/include/conf_keys.h` for reconfigure in this PR. It's
because `prefetch` is a parameter that's supported at both connection
and session configuration levels, and the current version of
`dist/api_config.py` is not able to generate more than 1 (e.g. both
connection and session level) config key and struct for the same
parameter, which will cause compilation failure in this case. The
solution I take here is to detect the duplicated config key and append a
suffix to the struct name and add it to the struct and keynumber lists.
To minimise impact this solution is applied to the "prefetch" config key
only for now. Though it's not an elegant solution it does the job.
Please let me know if you have better ideas how to address this issue.

Also I could not make the Catch2 unit test work in the new compiled
configuration setup, and plan to cover that in a separate ticket if it's
okay.