WT-15793 Fix creating empty tables concurrently with a checkpoint (#12580)
Fix the ability to create empty tables concurrently with a checkpoint.

The metadata of newly created tables are included in `copy_metadata_qh`.
The problem is that if this is done concurrently with a checkpoint, the
checkpoint processing will not copy the metadata entry (as it's
invisible to the checkpoint transaction) and then silently drop the
entry from the list. If the table remains empty, it will not be included
in the next checkpoint. Fix this issue by retrying the copy at the next
checkpoint.