WT-15168 Understand (but ignore) old configuration strings from metadata (#12256) (#12607)
This is pretty dumb right now.

I originally had some code to check that the values for these
configurations keys was empty, but the motivating problem was actually
old configs. That's more work to check because it's on the open path. A
check on session->create is probably acceptable, but I'm not sure that's
enough coverage to bother with. I'd like some feedback/discussion on
this decision.

Finally, I also sneaked in a couple of `self.assertEqual` spelling fixes
for newer Python versions.

(cherry picked from commit 3dee0127aac2131e8bdedef7f7c4fc1f6dd47310)

Summarize the reason behind this change (this might be the problem
you're solving, or the context around the request) and the solution you
have chosen.