WT-14648 Fix log subsystem returning EBUSY from conn->close() (#12050) (v8.0 backport) (#12610)
The investigation identified a real scenario where EBUSY could return
from conn->close(). The scenario is as such:
1. Log server is doing a log_printf() which requires a slot
2. Connection->close() needs to flush it's lsn and requires a slot to do
so
3. Connection->close() can't access the slot and returns EBUSY

This tickets makes sure that conn->close() does not return EBUSY
interacting with the log subsystem. The reproducer failing is on the
ticket.

**Update:** During this investigation i found two race condition bugs
wrlsn due to conn->close() shutting down. More informatino can be found
on the ticket.

(cherry picked from commit 4247228a571d7a813cf2f90020a63b8082e8e3c3)

Summarize the reason behind this change (this might be the problem
you're solving, or the context around the request) and the solution you
have chosen.

---------

Co-authored-by: jchen <29520270+jiechenbo@users.noreply.github.com>
Co-authored-by: Albert Song <albert.song@mongodb.com>