WT-15520 Dump the error log after a failed test/suite test (#12582)
After a failed test/suite test, dump the contents of the error log to
`stderr`. It skips calling the connection's event handler, as printing
directly to `stderr` results in an easier to read output.

The PR also overrides `wiredtiger_dump_error_log` in Python, so that we
can use a Python function as a callback.