WT-15743 Fix incorrect handling of WT_NOTFOUND when opening stable cursor (#12581)
To be honest, I'm not 100% comfortable with this. We still don't know
what conditions cause a `WT_NOTFOUND` rather than an `ENOENT` (see the
SLS ticket linked to the WT ticket), which means (1) it's hard to be
confident this fixes the root cause of the bug, and (2) we can't test
for it.

To deal with this, how about (in addition to this PR) I raise a new
ticket for investigating why we get `WT_NOTFOUND` and adding a test?