WT-15835 Fix two bugs in layered cursors, a casting error and a refinement for elegant step-down (#12620)
When opening subordinate cursors, it's possible that session->dhandle is
not the layered cursor. So fix a bug where it gets cast to a layered
cursor (we already have the layered cursor).

If we are stepping down, we can still continue if the next operation for
the cursor is an update, as long as all previous operations were not
updates. For elegant step-down, this allows more cases to proceed
without rollback.

---------

Co-authored-by: Don Anderson <donald.anderson@mongodb.com>
Co-authored-by: Ivan Kochin <ivan.kochin@mongodb.com>