WT-15843 change eviction update trigger to be equal to dirty by default (#12641)
There's a reoccurring issue from customers where application threads are
blocked from doing work due to update eviction trigger threshold being
exceeded. A common work around to this issue is to increase the setting
for the update trigger threshold. We can set the update threshold to the
20% default rather than 10% of cache to help alleviate this problem.
This PR changes the default for update trigger threshold to be equal to
the dirty trigger threshold when no other value has been specified.