WT-15826 disable building delta for in-memory split (#12669)
When we do an in-memory split and move an existing key from page A to B,
we don't mark it ineligible for deltas. so if it has some existing
deltas, when doing reconciliation, the deltas can be written to page A
instead of B and causes the same key to be duplicated in two pages. This
PR fixes the issue by disabling delta building for in-memory split.