WT-15892 Fix __wt_page_in_func segfault because use after free of the page (#12682)
The delta rewrite code changes the page inside ref underneath. While the
caller still has a pointer to the old page, which causes this issue.
Refresh the pointer in the caller to fix this.

This should all go away after we can generate a full page directly from
the base page and the deltas.