WT-14392 wt util no longer overwrites config when verifying metadata (#12702)
[BACKPORT-25602](https://jira.mongodb.org/browse/BACKPORT-25602)

…(#11910)

Prior to this change, the `wt` utility tool overwrote the provided
config string when also provided with the `-m` flag to verify metadata.

As per @kommiharibabu's suggestion - the metadata option `-m` now
utilises `metadata_config` instead of `cmd_config`. The code has also
been adjusted to account for this change.