WT-15536 Consider evict dirty pages if we cannot find enough clean pages (#12705)
When the clean page threshold is breached and insufficient clean pages
are available for eviction, the system may stall. Currently, dirty pages
are considered for eviction when the eviction process enters aggressive
mode. However, this approach may not suffice if eviction is unable to
make adequate progress. To mitigate this, dirty pages should also be
considered for eviction whenever the system cannot locate enough clean
pages for eviction.