WT-15951 Use weights for timing stress, which don't have to add up to 1 (#12714)
Introduce a new pair of macros to `test/model`, which allow the user to
easily initialize a list of weights that don't have to add up to 1. The
macros create a function for producing the sum of weights, which will
remain accurate even if the weights change after the initialization.

The reason for this macro is to ensure that there is only one list of
weights, instead of one list of weights when initializing them and one
list when adding them up, so that they would never get out of sync. If,
for example, someone added a new weight variable and initialized it, but
forgot to add it to where the weights are summed up, some options might
not ever be triggered (as explained on the ticket).

The macros were written to optimize the ease of use, not performance,
given that the callers to the created "total weights" function are not
on the critical path.