WT-14467 Modify for layered cursors (#12709)
This was quite a bit more involved than I thought -- the stuff about
`__cursor_needkey`, and notion of in-tree vs out-of-tree keys and values
is all new to me.

Still, there's a link to a patch build on the ticket -- I feel pretty
good about this, correctness-wise. It's definitely not as fast as it
could be. I think theoretically I should be able to do a bunch less
searching, and I possibly don't need to copy a full value from the
stable table when we can't find a base value in the ingest table. But
that's something we can worry about further down the line.

Testing-wise, this enables modifies in `test/checkpoint` and
`test/format`. There's also some new Python tests. There was only one
`cursor->modify` Python test previously, so to reuse the helpers in that
file I did some refactoring and put a lot of the heavy lifting in
`modify_utils.py`.