WT-15489 Write the oldest visible update we need to keep to the disk image for in-memory btree (#12764)
When reconciling an in-memory btree, we use the same eviction rules as
the non in-memory btree. This is not efficient especially when in-memory
btree is used in ingestable as we cannot overflow the updates to the
history store. Instead of writing the stable version to the disk image
and still keeping every version in memory on the update chain, we should
write the oldest version that we still need to keep to the disk image
and free it from the update chain.

This PR:

Splits __wti_rec_upd_select into separate paths for in-memory and
on-disk
In-memory trees now write the oldest update that we can keep. This can
be (ordered in terms of priority):
The first globally visible update
The oldest committed, non-tombstone update that is not pruned
Adds unit tests for update selection logic for refactoring purpose in
the future

---------

Co-authored-by: Chenhao Qu <chenhao.qu@mongodb.com>