WT-15184 Add more null pointer checks for avoiding Fault segment error (#12765)
WT-15184
There was an Atlas failure which hits a segmentation fault when trying
to dump each session's transaction state after eviction was stuck for
too long. We don't have a core dump to track down exactly which variable
caused the segmentation fault, but we can definitely add in some
conditions to check whether certain variables are NULL before attempting
to dereference them.