WT-15927 Enable unpacking of deltas and the base page during the merging process (#12731)
We currently use the base image to build the in-memory page, then apply
each delta to the update chain and call reconcile to write the new disk
image. The PR is part of merging base disk image and deltas into a new
disk image work for leaf page delta, initially we tried to unpack all
deltas first, and then do the merge. However it brings memory overhead
so this PR optimises this by unpacking the deltas only when necessary
during the k-way merging process.