WT-16050 Update the checkpoint generation before clearing the sync flag (#12791)
Previously, the checkpoint generation was updated after checkpointing
the btree. This introduced a timing gap during which eviction could
evict a page that should be included in the next checkpoint under
disaggregated storage—after the checkpoint was completed but before the
checkpoint generation was updated.

Additionally, we need to track the checkpoint generation for metadata,
as eviction must also avoid removing pages that do not belong to the
current metadata checkpoint.