Early return when one of constituents doesn't exist + position alternate basing on the current
Early return is important for the case when the case when a constituent hasn't been created
on the first operation, but then before the next, the creation occured so we still might need
to position the alternate cursor.

Positioning constituent should be done based on the current cursor position and not on the
iface cursor key position for the same reason.