WT-15856 Rework layered cursors positioning (#12825)
There is an issue with the current positioning logic for `next()` and
`prev()`. According to the comment in `__clayered_iterate_constituent`,
it expects either none or both of the constituents to be positioned.

However, there are cases where other functions that are supposed to
leave the cursor positioned end up returning only one cursor positioned
(which is also set as `current_cursor`). For example, this could happen
if `search()` finds the entry in the ingest table, so it doesn’t need to
access the stable table in this case. As a result, `stable_cursor` is
left unpositioned. This can lead to a situation where the cursor could
return out-of-order keys. Please refer to the information in the ticket
for more details.

This PR reworks the positioning logic to make it more straightforward
and better documented. It also adds documentation to establish a clear
contract between different interfaces regarding leaving a cursor
positioned.

Additionally, it introduces extra testing for layered cursors, which we
previously did not have, and moves the testing oplog implementation to
`helper_disagg.py` to allow sharing across multiple tests.