WT-15619 Implement rewriting leaf page deltas directly into a full page (#12804)
In disagg, currently if there's a base image with deltas, we need to
read the base image into memory, apply deltas to the in-memory page,
then call reconcile to build the full page. This ticket is to implement
rewriting leaf page deltas directly into a full disk image to improve
the performance of loading page into memory.