WT-14586 Handle the scenario of server crash during table creation (#12808)
A crash that happen during schema operations on a logged table can leave
WiredTiger metadata table in an inconsistent state. This is because all
schema metadata operations are non-transactional, and thus the log
records each metadata operation in it's own commit.

Therefore it is possible after log replay that the metadata table has
incomplete tables with possibly no file: or colgroup: metadat entries
because we crashed at a wrong time. The ticket here reads through the
metadata file and fixes up any incomplete simple tables.