WT-16201 Additional logging of checkpoint pickup (#12890)
We want to make sure there is enough logging at verbose level 1 to know
where time is being spent during checkpoint pickup. But we want this to
be light enough that we could, potentially, have it on full time. I.e.,
the logging should be fixed per checkpoint and not proportional to
number of entries in the metadata.

This change:

- Includes a message at the beginning of each of the 3 steps of
checkpoint pickup (see comments in `__disagg_pick_up_checkpoint`).
- Prints a summary of the table processing, the number of existing
tables vs. new tables processed. Taking 15 seconds to process a
checkpoint with a million tables is different than taking 15 seconds to
process a checkpoint with just one table.
- Includes a message at the end of checkpoint pickup.

All at verbose level 1.

My testing has been minimal. It compiles and passes the layered table
python tests.