WT-15827 Catch "buffer allocated and never discarded" in wt tool dump_offsets (#12912)
Introduced `WT_USE_ERR` macro to capture the return value of `usage`
function call and `goto err` so that allocated resources are properly
cleaned up.

Also introduced a python test to ensure we no longer encounter this
leak. As a sanity check, I ran the test without the fix to ensure it
fails appropriately, and it does.

---------

Co-authored-by: Etienne Petrel <77254506+etienneptl@users.noreply.github.com>