WT-16167 test/format collection hashing (#12904)
This calculates a checksum over the database and, for now, simply
reports it.

It uses an FNV-1A hash, since that's simple and works for incremental
hashing (to avoid loading all keys/values into memory somehow).

There is a little bit of finagling with the keys, to ensure that hashes
are comparable across tables that happen to use different prefixes.

Checksum comparison across nodes is coming in a future ticket.