WT-15949 Implement bitflip detection on checksum mismatch (#12928)
This change closes a gap we have w.r.t. AF corruption analysis.
Essentially we flip every bit in the block and recompute the checksum.
If it matches we log success and the location of the bit, if it doesn't
we log that we attempted to do the bitflip detection but failed.

Comes with unit testing as python testing is painfully hard.