WT-16054 Implement KEK write PALI path mechanism (#12926)
Originally this ticket had the full write path mechanism completed but
it was too big of a PR. I split it into two PRs. I intend to do the
second part and write extensive testing in WT-16290 (pt.2).

This ticket will focus on writing the PALI mechanism to the new key
provider table with no header. This ticket plans to establish a write
mechanism to PALI as such:
- WT calls the new API get_key_blob to retrieve the current KEK blob. If
the call indicates that no changes were made, we skip early.
- WT calls the get_complete API to return the LSN to the encryption
layer, and to discard the memory used to store the blob.

**Note: There is minimal testing for this PR. I plan to write extensive
testing in WT-16290.**
The work on adding the header will be done in WT-16298.