WT-16263 Add disagg address cookie decode tooling (#12932)
I've added a new class in btree_format that decodes the disagg address
cookie. This is then used in `wt_binary_decode` when decoding row
address cells, as well as in a new script `wt_disagg_addr_decode` that
accepts a hex string from the command line. The decoded cells will allow
developers to find child pages of internal pages, and the cli tool can
be used to find the root page metadata from the checkpoint string in the
metadata page/table.