WT-14115 add stat for file reuseable portation (#12973)
Add two new stats to indicate the count of data files which has 50%
reusable space and 90% reusable space. The target of the stats is to
improve observability for fragment related issues.
A new test case is introduced to validate these stats. As block size
only works with ASC mode, so the test case skips `disagg` and `tiered`
hooks.

---------

Co-authored-by: Albert Song <albert.song@mongodb.com>
Co-authored-by: Etienne Petrel <77254506+etienneptl@users.noreply.github.com>